Add bulk submission method for trusted users
