+++ devvit.yaml
@@ -1,2 +1,2 @@
 name: bot-bouncer
-version: 1.2.14
+version: 1.2.16
